+++ adapt/src/main/java/io/github/vshnv/adapt/dsl/BindScope.kt
@@ -1,4 +1,4 @@
-package io.github.vshnv.adapt
+package io.github.vshnv.adapt.dsl
 
 import androidx.recyclerview.widget.RecyclerView
