+++ adapt/src/main/java/io/github/vshnv/adapt/dsl/Bindable.kt
@@ -1,4 +1,4 @@
-package io.github.vshnv.adapt
+package io.github.vshnv.adapt.dsl
 
 interface Bindable<T, V> {
     fun bind(bindView: BindScope<T, V>.() -> Unit): LifecycleRenewAttachable<T, V>
